UPDATE Product-sites
Megcsináltam a paneleket a főoldalon

+++ Projekt/webshop-app/database/migrations/2024_11_05_101002_create_products_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string("name");
             $table->text("description");
-            $table->decimal("price");
+            $table->integer("price");
             $table->integer("category_id");
             $table->integer("stock");
             $table->string("image");
